+++ streamingpro-spark-2.0/src/main/java/streaming/dsl/mmlib/algs/SQLPython.scala
@@ -1,0 +1,8 @@
+package streaming.dsl.mmlib.algs
+
+/**
+  * Created by allwefantasy on 1/2/2018.
+  */
+class SQLPython {
+
+}
